Sprout - Never talk to disabled providers

+++ sprout/appliances/tasks.py
@@ -183,6 +183,8 @@ def kill_appliance_delete(self, appliance_id, _delete_already_issued=False):
     delete_issued = False
     try:
         appliance = Appliance.objects.get(id=appliance_id)
+        if not appliance.provider.is_working:
+            raise RuntimeError('Provider is not working.')
         if appliance.provider_api.does_vm_exist(appliance.name):
             appliance.set_status("Deleting the appliance from provider")
             # If we haven't issued the delete order, do it now
@@ -318,6 +320,8 @@ def poke_trackerbot(self):
     # Similarly if some of them becomes usable ...
     for provider_id, template_name, usability in template_usability:
         provider, create = Provider.objects.get_or_create(id=provider_id)
+        if not provider.working or provider.disabled:
+            continue
         with transaction.atomic():
             for template in Template.objects.filter(provider=provider, original_name=template_name):
                 template.usable = usability
@@ -338,7 +342,7 @@ def create_appliance_template(self, provider_id, group_id, template_name, source
     """This task creates a template from a fresh CFME template. In case of fatal error during the
     operation, the template object is deleted to make sure the operation will be retried next time
     when poke_trackerbot runs."""
-    provider = Provider.objects.get(id=provider_id)
+    provider = Provider.objects.get(id=provider_id, working=True, disabled=False)
     provider.cleanup()  # Precaution
     group = Group.objects.get(id=group_id)
     with transaction.atomic():
@@ -407,6 +411,8 @@ def create_appliance_template(self, provider_id, group_id, template_name, source
 def prepare_template_deploy(self, template_id):
     template = Template.objects.get(id=template_id)
     try:
+        if not template.provider.is_working:
+            raise RuntimeError('Provider is not working.')
         if not template.exists_in_provider:
             template.set_status("Deploying the template.")
             provider_data = template.provider.provider_data
@@ -514,8 +520,10 @@ def prepare_template_seal(self, template_id):
 @singleton_task()
 def prepare_template_poweroff(self, template_id):
     template = Template.objects.get(id=template_id)
-    template.set_status("Powering off")
     try:
+        if not template.provider.is_working:
+            raise RuntimeError('Provider is not working.')
+        template.set_status("Powering off")
         template.provider_api.stop_vm(template.name)
         template.provider_api.wait_vm_stopped(template.name)
     except Exception as e:
@@ -528,8 +536,10 @@ def prepare_template_poweroff(self, template_id):
 @singleton_task()
 def prepare_template_finish(self, template_id):
     template = Template.objects.get(id=template_id)
-    template.set_status("Finishing template creation.")
     try:
+        if not template.provider.is_working:
+            raise RuntimeError('Provider is not working.')
+        template.set_status("Finishing template creation.")
         if template.temporary_name is None:
             tmp_name = "templatize_{}".format(fauxfactory.gen_alphanumeric(8))
             Template.objects.get(id=template_id).temporary_name = tmp_name  # metadata, autosave
@@ -556,8 +566,10 @@ def prepare_template_delete_on_error(self, template_id):
         template = Template.objects.get(id=template_id)
     except ObjectDoesNotExist:
         return True
-    template.set_status("Template creation failed. Deleting it.")
     try:
+        if not template.provider.is_working:
+            raise RuntimeError('Provider is not working.')
+        template.set_status("Template creation failed. Deleting it.")
         if template.provider_api.does_vm_exist(template.name):
             template.provider_api.delete_vm(template.name)
             wait_for(template.provider_api.does_vm_exist, [template.name], timeout='5m', delay=10)
@@ -776,8 +788,10 @@ def clone_template_to_appliance__clone_template(self, appliance_id, lease_time_m
             self.request.callbacks[:] = []
             kill_appliance.delay(appliance_id)
             return
-    appliance.provider.cleanup()
+    if not appliance.provider.is_working:
+        raise RuntimeError('Provider is not working.')
     try:
+        appliance.provider.cleanup()
         if not appliance.provider_api.does_vm_exist(appliance.name):
             appliance.set_status("Beginning template clone.")
             provider_data = appliance.template.provider.provider_data
@@ -846,6 +860,8 @@ def clone_template_to_appliance__wait_present(self, appliance_id):
         # source objects are not present, terminating the chain
         self.request.callbacks[:] = []
         return
+    if not appliance.provider.is_working:
+        raise RuntimeError('Provider is not working.')
     if appliance.appliance_pool is not None:
         if appliance.appliance_pool.not_needed_anymore:
             # Terminate task chain
@@ -882,6 +898,8 @@ def appliance_power_on(self, appliance_id):
         # source objects are not present
         return
     try:
+        if not appliance.provider.is_working:
+            raise RuntimeError('Provider is not working.')
         if appliance.provider_api.is_vm_running(appliance.name):
             try:
                 current_ip = appliance.provider_api.current_ip_address(appliance.name)
@@ -956,6 +974,8 @@ def appliance_power_off(self, appliance_id):
         # source objects are not present
         return
     try:
+        if not appliance.provider.is_working:
+            raise RuntimeError('Provider is not working.')
         if appliance.provider_api.is_vm_stopped(appliance.name):
             Appliance.objects.get(id=appliance_id).set_status("Appliance was powered off.")
             with transaction.atomic():
@@ -990,6 +1010,8 @@ def appliance_suspend(self, appliance_id):
         # source objects are not present
         return
     try:
+        if not appliance.provider.is_working:
+            raise RuntimeError('Provider is not working.')
         if appliance.provider_api.is_vm_suspended(appliance.name):
             Appliance.objects.get(id=appliance_id).set_status("Appliance was suspended.")
             with transaction.atomic():
@@ -1017,6 +1039,8 @@ def retrieve_appliance_ip(self, appliance_id):
     """Updates appliance's IP address."""
     try:
         appliance = Appliance.objects.get(id=appliance_id)
+        if not appliance.provider.is_working:
+            raise RuntimeError('Provider is not working.')
         appliance.set_status("Retrieving IP address.")
         ip_address = appliance.provider_api.current_ip_address(appliance.name)
         if ip_address is None:
@@ -1046,7 +1070,7 @@ def refresh_appliances_provider(self, provider_id):
     appliances stored in database.
     """
     self.logger.info("Refreshing appliances in {}".format(provider_id))
-    provider = Provider.objects.get(id=provider_id)
+    provider = Provider.objects.get(id=provider_id, working=True, disabled=False)
     if not hasattr(provider.api, "all_vms"):
         # Ignore this provider
         return
@@ -1085,18 +1109,18 @@ def refresh_appliances_provider(self, provider_id):
 @singleton_task()
 def check_templates(self):
     self.logger.info("Initiated a periodic template check")
-    for provider in Provider.objects.all():
+    for provider in Provider.objects.filter(disabled=False):
         check_templates_in_provider.delay(provider.id)
 
 
 @singleton_task(soft_time_limit=180)
 def check_templates_in_provider(self, provider_id):
     self.logger.info("Initiated a periodic template check for {}".format(provider_id))
-    provider = Provider.objects.get(id=provider_id)
+    provider = Provider.objects.get(id=provider_id, disabled=False)
     # Get templates and update metadata
     try:
         templates = map(str, provider.api.list_template())
-    except:
+    except Exception:
         provider.working = False
         provider.save()
     else:
@@ -1310,31 +1334,33 @@ def wait_appliance_ready(self, appliance_id):
 
 @singleton_task()
 def anyvm_power_on(self, provider, vm):
-    provider = Provider.objects.get(id=provider)
+    provider = Provider.objects.get(id=provider, working=True, disabled=False)
     provider.api.start_vm(vm)
 
 
 @singleton_task()
 def anyvm_power_off(self, provider, vm):
-    provider = Provider.objects.get(id=provider)
+    provider = Provider.objects.get(id=provider, working=True, disabled=False)
     provider.api.stop_vm(vm)
 
 
 @singleton_task()
 def anyvm_suspend(self, provider, vm):
-    provider = Provider.objects.get(id=provider)
+    provider = Provider.objects.get(id=provider, working=True, disabled=False)
     provider.api.suspend_vm(vm)
 
 
 @singleton_task()
 def anyvm_delete(self, provider, vm):
-    provider = Provider.objects.get(id=provider)
+    provider = Provider.objects.get(id=provider, working=True, disabled=False)
     provider.api.delete_vm(vm)
 
 
 @singleton_task()
 def delete_template_from_provider(self, template_id):
     template = Template.objects.get(id=template_id)
+    if not template.provider.is_working:
+        raise RuntimeError('Provider is not working.')
     try:
         template.provider_api.delete_template(template.name)
     except Exception as e:
@@ -1353,6 +1379,8 @@ def appliance_rename(self, appliance_id, new_name):
         appliance = Appliance.objects.get(id=appliance_id)
     except ObjectDoesNotExist:
         return None
+    if not appliance.provider.is_working:
+        raise RuntimeError('Provider is not working.')
     if not appliance.provider.allow_renaming:
         return None
     if appliance.name == new_name:
@@ -1379,6 +1407,8 @@ def rename_appliances_for_pool(self, pool_id):
         ]
         for appliance in appliances:
             if not appliance.provider.allow_renaming:
+                continue
+            if not appliance.provider.is_working:
                 continue
             new_name = '{}_'.format(appliance_pool.owner.username)
             if appliance.version and not appliance.version.startswith('...'):
@@ -1443,7 +1473,7 @@ def calculate_provider_management_usage(self, appliance_ids):
             if provider_key not in results:
                 results[provider_key] = []
             results[provider_key].append(appliance.id)
-    for provider in Provider.objects.all():
+    for provider in Provider.objects.filter(working=True, disabled=False):
         provider.appliances_manage_this_provider = results.get(provider.id, [])
 
 
@@ -1505,6 +1535,8 @@ def obsolete_template_deleter(self):
 @singleton_task()
 def connect_direct_lun(self, appliance_id):
     appliance = Appliance.objects.get(id=appliance_id)
+    if not appliance.provider.is_working:
+        raise RuntimeError('Provider is not working.')
     if not hasattr(appliance.provider_api, "connect_direct_lun_to_appliance"):
         return False
     try:
@@ -1523,6 +1555,8 @@ def connect_direct_lun(self, appliance_id):
 @singleton_task()
 def disconnect_direct_lun(self, appliance_id):
     appliance = Appliance.objects.get(id=appliance_id)
+    if not appliance.provider.is_working:
+        raise RuntimeError('Provider is not working.')
     if not appliance.lun_disk_connected:
         return False
     if not hasattr(appliance.provider_api, "connect_direct_lun_to_appliance"):
@@ -1718,7 +1752,7 @@ def parsedate(d):
 @singleton_task()
 def synchronize_untracked_vms_in_provider(self, provider_id):
     """'re'-synchronizes any vms that might be lost during outages."""
-    provider = Provider.objects.get(id=provider_id)
+    provider = Provider.objects.get(id=provider_id, working=True, disabled=False)
     provider_api = provider.api
     if not hasattr(provider_api, 'list_vm'):
         # This provider does not have VMs (eg. Hawkular or Openshift)
@@ -1892,7 +1926,9 @@ def process_docker_images_from_url_group(self, group_id, version, docker_version
         date = datetime.strptime(date, '%Y%m%d%H%M%S').date()  # noqa
     except AttributeError:
         raise ValueError('Could not parse date from {}'.format(docker_version))
-    for provider in Provider.objects.exclude(container_base_template=None):
+    for provider in Provider.objects.exclude(
+            Q(container_base_template=None) | Q(container_base_template='')).filter(
+                working=True, disabled=False):
         try:
             Template.objects.get(
                 ~Q(container=None), template_group=group, provider=provider, version=version,
@@ -1911,7 +1947,7 @@ def docker_vm_name(version, date):
 @singleton_task()
 def create_docker_vm(self, group_id, provider_id, version, date, pull_url):
     group = Group.objects.get(id=group_id)
-    provider = Provider.objects.get(id=provider_id)
+    provider = Provider.objects.get(id=provider_id, working=True, disabled=False)
     with transaction.atomic():
         if provider.remaining_configuring_slots < 1:
             self.retry(
@@ -1958,7 +1994,7 @@ def sync_appliance_hw(self, appliance_id):
 
 @singleton_task()
 def sync_provider_hw(self, provider_id):
-    Provider.objects.get(id=provider_id).perf_sync()
+    Provider.objects.get(id=provider_id, working=True, disabled=False).perf_sync()
 
 
 @singleton_task()
